\name{KappaV-package}
\alias{KappaV-package}
\docType{package}
\title{
Calculates vectorial Kappa, i.e. congruence between two vectorial landscapes.

}
\description{
KappaV allows to calculate vectorial Kappa (Kv), that extends the principle of Cohen's Kappa index in spatial ecology. To assess this congruence, Kv deals with areas of intersected patches between two landscapes. It thus provides an exact alternative of the Kappa rather than rasterizing and then calculating the classical Cohen's Kappa calculation when working on vectorial landscapes.
}
\details{
\tabular{ll}{
Package: \tab KappaV\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-31\cr
License: \tab GPL(>=2)\cr
URL: \tab http://www.vincentbonhomme.fr/KappaV \cr
Depends: \tab maptools, sp, rgeos, PresenceAbsence, R(>= 2.15.0) \cr
}

\bold{Have a look to the package's vignette and to the package website: \url{http://www.vincentbonhomme.fr/KappaV}}
}
\author{
Vincent Bonhomme

Maintainer: Vincent Bonhomme <bonhomme.vincent@gmail.com>
}
\references{
We are working on the companion paper.
}

\examples{
# have a look to the package's vignette: http://www.vincentbonhomme.fr/KappaV
}
